+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ### Removed
 
 - robots.txt not found error in lods
+- Duplicate logs (_logs service)